added XmlParser code to Wordpress.php
